requirements.txt
# Generated requirements.txt from project imports
# Run this script to update: python sript.py

exchange_calendars
loguru==0.7.3
numpy==2.2.3
pandas==2.3.0
pkg_resources
requests==2.32.3
scipy==1.13.1
setuptools==75.1.0
tzlocal==5.3.1
PyYAML==6.0.2
